feat: move range_planning into analysis and update the docs and tests
