+++ docs/api/analysis/range_planning.md
@@ -1,0 +1,3 @@
+# Range Planning
+
+::: pyretailscience.analysis.range_planning
