--- docs/api/range_planning.md
@@ -1,3 +1,0 @@
-# Range Planning
-
-::: pyretailscience.range_planning
